Inserting Details page inside modal
