+++ web/src/constants/actionTypes.js
@@ -17,22 +17,29 @@ export const GET_ALL_SEASONS_SUCCESS = "GET_ALL_SEASONS_SUCCESS";
 export const GET_ALL_SEASONS_FAILED = "GET_ALL_SEASONS_FAILED";
 export const SET_ACTIVE_SEASON = "SET_ACTIVE_SEASON";
 export const SET_ACTIVE_SEASON_FAILED = "SET_ACTIVE_SEASON_FAILED";
+
 export const SET_CHALLENGES_LIST = "SET_CHALLENGES_LIST";
 export const SET_CHALLENGES_LIST_FAILED = "SET_CHALLENGES_LIST_FAILED";
+
 export const FETCH_PREFERENCES_SUCCESS = "FETCH_PREFERENCES_SUCCESS";
 export const FETCH_PREFERENCES_FAILED = "FETCH_PREFERENCES_FAILED";
 export const GET_ALL_SEASON_TEAMS_SUCCESS = "GET_ALL_SEASON_TEAMS_SUCCESS";
 export const GET_ALL_SEASON_TEAMS_FAILED = "GET_ALL_SEASON_TEAMS_FAILED";
+
 export const GET_CHALLENGE_DETAILS_SUCCESS = "GET_CHALLENGE_DETAILS_SUCCESS";
 export const GET_CHALLENGE_DETAILS_FAILED = "GET_CHALLENGE_DETAILS_FAILED";
+export const CLEAN_CHALLENGE_DETAIL = "CLEAN_CHALLENGE_DETAIL";
+
 export const GET_TEAM_DETAILS_SUCCESS = "GET_TEAM_DETAILS_SUCCESS";
 export const GET_TEAM_DETAILS_FAILED = "GET_TEAM_DETAILS_FAILED";
+
 export const BUY_CHALLENGE_SUCCESS = "BUY_CHALLENGE_SUCCESS";
 export const BUY_CHALLENGE_FAILED = "BUY_CHALLENGE_FAILED";
 export const VALIDATE_CHALLENGE_SUCCESS = "VALIDATE_CHALLENGE_SUCCESS";
 export const VALIDATE_CHALLENGE_FAILED = "VALIDATE_CHALLENGE_FAILED";
 export const CLOSE_CHALLENGE_SUCCESS = "CLOSE_CHALLENGE_SUCCESS";
 export const CLOSE_CHALLENGE_FAILED = "CLOSE_CHALLENGE_FAILED";
+
 export const SET_ACTIVE_TEAM = "SET_ACTIVE_TEAM";
 export const CREATE_TEAM_SUCCESS = "CREATE_TEAM_SUCCESS";
 export const CREATE_TEAM_FAILED = "CREATE_TEAM_FAILED";
